Added Simple Authentication to resolve #6
