+++ simple-auth/src/main/java/com/fga/example/config/OpenFgaProperties.java
