+++ simple-auth/src/main/java/com/fga/example/controller/DocumentController.java
