+++ simple-auth/src/main/java/com/fga/example/fga/FgaCheck.java
