+++ simple-auth/src/main/java/com/fga/example/fga/OpenFga.java
