+++ simple-auth/src/main/java/com/fga/example/service/DocumentService.java
