+++ simple-auth/src/main/resources/example-auth-model.json
